fix(PST)
When the ArNS Service is down, fetching the contract via Warp can take up to 2 minutes. Currently, Warp is also failing, leading to a worst-case scenario of 4 minutes before an upload starts. This happens because contract data is fetched in two separate steps, and if the first attempt fails, the second attempt retries, further delaying the upload process.

Solution

This PR introduces error caching to optimize the upload preparation phase. If fetching the community contract fails, the error will be cached for 30 minutes, preventing redundant fetch attempts and significantly reducing the time spent waiting before an upload starts.